feat(web): Split up PlannerFactoryProducts into their subcomponents
